refactor: migrate CLI commands from shell to cli package
This commit moves the CLI command implementations from the shell package to a new cli package. Changes include:

- Create new cli package with database, gui, plugin, and version commands
- Update main.go to import and use the new cli package
- Maintain existing command functionality and structure
- Add .gitkeep files for new directories
