--- export.html
@@ -1,37 +1,0 @@
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Export</title>
-</head>
-<body>
-    <table>
-        <thead>
-            <tr>
-                <th>Name</th>
-                <th>Email</th>
-                <th>Birthday</th>
-            </tr>
-        </thead>
-        <tbody>
-            <tr>
-                <td>John Doe</td>
-                <td>johndoe@example.com</td>
-                <td>1990-01-01T00:00:00Z</td>
-            </tr>
-            <tr>
-                <td>Jane Doe</td>
-                <td>janedoe@example.com</td>
-                <td>1995-01-01T00:00:00Z</td>
-            </tr>
-            <tr>
-                <td>Bob Smith</td>
-                <td>bobsmith@example.com</td>
-                <td>1980-01-01T00:00:00Z</td>
-            </tr>
-        </tbody>
-    </table>
-</body>
-</html>
