+++ test/mocks/stringer_mock.go
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 The PolyClient Authors
+//
+// SPDX-License-Identifier: GPL-3.0-or-later WITH LicenseRef-PolyClient-Plugin-Exception
+
 package mocks
 
 import "fmt"
